improvements done just missing manage files as variables
