--- src/test/kotlin/test/common/LoginUser.kt
@@ -1,3 +1,0 @@
-package test.common
-
-data class LoginUser(val login: String, val password: String)
