--- src/test/kotlin/test/common/ScenarioContext.kt
@@ -1,8 +1,0 @@
-package test.common
-
-import org.springframework.web.client.RestTemplate
-
-class ScenarioContext {
-    val restTemplate = RestTemplate()
-
-}
