+++ src/test/kotlin/test/common/SubmitFeatureContext.kt
@@ -1,12 +1,15 @@
 package test.common
 
-import java.net.PasswordAuthentication
+import org.springframework.http.HttpHeaders
+import org.springframework.util.LinkedMultiValueMap
+import org.springframework.web.client.RestTemplate
+import java.nio.file.Files
 
-class SubmitFeatureContext {
+object SubmitFeatureContext {
 
-    lateinit var urlEnvironment: String
-    lateinit var ftpUrl: String
-    lateinit var userName: String
-    lateinit var password: String
-
+    val restTemplate = RestTemplate()
+    val variables: MutableMap<String, String> = mutableMapOf()
+    val tempFile = Files.createTempDirectory("tempFolder").toFile()
+    var formDataBodyRequest = LinkedMultiValueMap<String, Any>()
+    val headers = HttpHeaders()
 }
